backend/requirements.txt
fastapi
uvicorn[standard]
httpx
requests
python-multipart
python-dotenv
playwright
pytesseract
pillow
numpy
opencv-python-headless
scikit-image
